requirements.txt
torch
torchvision
torchaudio
numpy
matplotlib
pillow
mitdeeplearning
streamlit
